move rollouts api stubs and "helpers"
