+++ api/stubs/rollouts/v1alpha1/analysis_types.go
