+++ api/stubs/rollouts/v1alpha1/groupversion_info.go
