+++ api/stubs/rollouts/v1alpha1/types.go
