+++ api/stubs/rollouts/v1alpha1/zz_generated.deepcopy.go
